(GH-43) Move extensions to domain namespace
